+++ packages/grafana-ui/src/components/TimePicker/TimeRangePicker.test.tsx
@@ -1,8 +1,8 @@
 import React from 'react';
 import { shallow } from 'enzyme';
-import { UnthemedTimePicker } from './TimePicker';
+import { UnthemedTimeRangePicker } from './TimeRangePicker';
 import { dateTime, TimeRange } from '@grafana/data';
-import dark from './../../themes/dark';
+import dark from '../../themes/dark';
 
 const from = '2019-12-17T07:48:27.433Z';
 const to = '2019-12-18T07:48:27.433Z';
@@ -16,7 +16,7 @@ const value: TimeRange = {
 describe('TimePicker', () => {
   it('renders buttons correctly', () => {
     const wrapper = shallow(
-      <UnthemedTimePicker
+      <UnthemedTimeRangePicker
         onChange={value => {}}
         value={value}
         onMoveBackward={() => {}}
@@ -30,7 +30,7 @@ describe('TimePicker', () => {
 
   it('renders content correctly after beeing open', () => {
     const wrapper = shallow(
-      <UnthemedTimePicker
+      <UnthemedTimeRangePicker
         onChange={value => {}}
         value={value}
         onMoveBackward={() => {}}
